Move radeon-profile-daemon-server socket file into its own subdir under /run
The rationale here is that QT has no methods to run chown on some
files but we do not want the socket being world readable. So instead of
writing lenghty C++-code using chown, we simply move the socket file into a
subdir which we can put our preferred permissions on during startup of the
daemon.

Signed-off-by: Lars Wendler <polynomial-c@gentoo.org>